Create README.md
MY DIGITAL KEYS - TERMS OF USE
Article 1 – Scope of application of the present terms of use
The present terms of use regulate the procedure made available by the Federal Government
for the electronic registration, identification and authentication of end users, who may or may 
not be citizens. Using this procedure enables end users to register with a view to gaining safe 
access to electronic services provided by the various Belgian authorities and safe electronic
communications with the said authorities.
Nonetheless, it may be that some public bodies have different systems in place for electronic
end user management.
Article 2 - Access to the procedure
The end user has access to the procedure, albeit without any guarantees being offered that 
access to the procedure and the services made available shall be guaranteed at all times or be 
free from errors or technical disruptions. 
Access to the procedure may be disconnected at any point in time, in full or in part, (for 
maintenance purposes, in amongst other things). Insofar as reasonably possible, the end user
shall be notified of such interruptions ahead of time.
The end user shall have access to certain services made available by the various Belgian 
authorities only upon completing the applicable registration, identification and authentication 
procedure. 
In doing so, the end user shall be required to: 
agree to the present terms of use; 
provide a working e-mail address. 
If necessary, the end user shall be required to amend the details relating to him/her to ensure 
they are up to date and accurate at all times.
Article 3 - Use of digital keys
The end user’s access to specific services, which are made available electronically, requires the 
use of digital keys (eID card reader, wireless eID card reader, security code by mobile app/text 
message/token and username and password, …). 
These digital keys and the details associated therewith are strictly personal and non￾transferable. 
Each end user is responsible for the safekeeping, protection, confidentiality and management 
of his/her digital keys and the details associated therewith. 
The end user is responsible for choosing a safe password or other secret code. 
If an end user is aware that his/her username, password, token or other digital key has been 
lost, or is being unlawfully used by third parties, or suspects such loss or unlawful use, he/she 
is to promptly put in place all relevant steps to deactivate the digital key as set out in article 6
and elsewhere.
In the event his/her digital key should be locked, the end user is to apply for a new digital key.